pyproject.toml
[build-system]
requires    = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ttt-jailbreak"
version = "0.1.0"
description = "Test-time training (TTT) jailbreak attacks against safety-aligned LLMs"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.11,<3.13"
dependencies = [
    "accelerate>=0.27.0",
    "aiohttp>=3.9.0",
    "datasets>=2.18.0",
    "fastapi>=0.110.0",
    "gitpython>=3.1.45",
    "hf-transfer>=0.1.9",
    "hydra-core>=1.3.2",
    "jsonlines>=4.0.0",
    "nltk>=3.9.2",
    "numpy>=1.24.0",
    "peft>=0.10.0",
    "pyyaml>=6.0",
    "scikit-learn>=1.7.2",
    "tabulate>=0.9.0",
    "torch>=2.0.0",
    "tqdm>=4.66.0",
    "transformers>=4.40.0,<5",
    "uvicorn>=0.27.0",
    "vllm==0.11.0",
]

[project.optional-dependencies]
tinker = [
    "tinker>=0.1.0",
    "tinker-cookbook>=0.4.1",
]

[tool.setuptools.packages.find]
where    = ["src"]
include  = ["ttt_jailbreak*"]

[tool.setuptools.package-dir]
"" = "src"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
